Create Challenge 1 code
Following the same challenge, I've also shared the code I developed to make the web page  works. Since it's my first experience on coding. It was really intensive and demanding.  There were a lot of questions and concerns that were able to be solved thanks for the DIO classes, and specially the blog, such as 'w3schools' which fundamentaly important to conclude this challenge. 
Most of all, I'm extremely happy to do it without requiring AI assistance and to push myself to do it 100% in English. Let's keep up the hard work!.