Add files via upload
My very first attempt to conclude DIO challenge. I was required to build my first complete website by applying forms, tables, midias and contact. This website is indeed to be for a health clinic (fake) so we could feel some real life experience on how to get the job done. I took me around 8 hours to complete this task, and the most difficult step was dealing with tables. I couldn't figure out how to attach it before the footer section. Overall all links work properly and the navigation is flowy.